LayoutTests/fast/dom/Element/getClientRects-return-type.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description("Tests that Element.getClientRects() returns a list of DOMRect objects.");

const rects = document.body.getClientRects();
shouldBe("rects.length", "1");
shouldBe("rects[0].__proto__", "DOMRect.prototype");
shouldBe("rects[0]", "rects.item(0)");
</script>
<script src="../../../resources/js-test-post.js"></script>>
</body>
</html>
